fixed test

+++ src/test/java/com/epam/training/auction/common/BiddingRaceExceptionTest.java
@@ -10,7 +10,7 @@ public class BiddingRaceExceptionTest {
     @Test
     public void checkExists(){
         UserTransferObject user = new UserTransferObject(1, "user", "pass");
-        double otherBid = 10.0;
+        long otherBid = 10;
         BiddingRaceException target = new BiddingRaceException(user, otherBid);
         assertEquals("User was not set on the exception.", user, target.getOtherBidder());
         assertTrue("Other Bid was not sett on the exception.", otherBid == target.getOtherBid());
